// frontend/src/app/modules/client/core/model/student/student-create.model.ts
export class StudentCreateModel {

  first_name: string;
  last_name: string;

  address: string;

  father_name: string;
  mother_name: string;
  father_contact: string;
  father_occupation_id: number;
  mother_occupation_id: number;
  mother_contact: string;

  guardian_name: string;
  guardian_contact: string;

  gender: number;
  date_of_birth: Date;

  class_id: number;
  section_id: number;
  roll_number: string;
  registration_number: string;

  bus_route_id: number;

  weight: number;
  height: number;
  nationality_id: number;
  blood_group: number;
}